Code Refactor and Addition of Feature
- Minor optimisations
- Resolved #13
- Separated the Cobra CMDs into their own files
- Added a .bat file to compile the program
