+++ src/versioninfo.rc
@@ -1,0 +1,30 @@
+1 VERSIONINFO
+FILEVERSION     1,1,0,0
+PRODUCTVERSION  1,1,0,0
+FILEFLAGSMASK   0X3FL
+FILEFLAGS       0L
+FILEOS          0X40004L
+FILETYPE        0X1
+FILESUBTYPE     0
+BEGIN
+    BLOCK "StringFileInfo"
+    BEGIN
+        BLOCK "040904B0"
+        BEGIN
+		    VALUE "CompanyName", ""
+            VALUE "FileDescription", "CLI tool of Cultured Downloader"
+            VALUE "FileVersion", "1.1.0"
+            VALUE "InternalName", "Cultured Downloader CLI"
+            VALUE "LegalCopyright", "KJHJason. GPL-3.0 license"
+            VALUE "OriginalFilename", "cultured-downloader-cli.exe"
+            VALUE "ProductName", "Cultured Downloader CLI"
+			VALUE "ProductVersion", "1.1.0"
+        END
+    END
+    BLOCK "VarFileInfo"
+    BEGIN
+            VALUE "Translation", 0x0409, 0x04B0
+    END
+END
+
+1 ICON "../res/cultured_downloader.ico"
